+++ database/migrations/2024_05_07_134028_create_credits_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('credits', function (Blueprint $table) {
             $table->id();
-            $table->bigInteger('amount');
+            $table->decimal('amount',20);
             $table->unsignedBigInteger('customer_id');
             $table->unsignedBigInteger('invoice_id')->nullable();
             $table->text('data')->nullable()->default(null);
